+++ web/twa-vis-framework/library/src/ts/io/data_store.ts
@@ -63,19 +63,19 @@ class DataStore {
      * 
      * @returns Promise that fulfills when all loading is complete
      */
-    public loadDataGroups(dataJSON: Object, value: number) {
+    public loadDataGroups(dataJSON: Object, dimension: string, value: number) {
         if(dataJSON == null) {
             let self = this;
             
             return $.getJSON("./data.json", function(json) {
-                self.recurseLoadDataGroups(json, null, null, self.dataGroups.length, value);
+                self.recurseLoadDataGroups(json, null, null, self.dataGroups.length, dimension, value);
             }).fail((error) => {
                 throw error;
             });    
         }
 
         return new Promise((resolve, reject) => {
-            this.recurseLoadDataGroups(dataJSON, null, null, this.dataGroups.length, value);
+            this.recurseLoadDataGroups(dataJSON, null, null, this.dataGroups.length, dimension, value);
             resolve("done");
         });
     }
@@ -84,7 +84,7 @@ class DataStore {
      * Recursively parses the visualisation definition file into hierarchal
      * DataGroup instances.
      */
-    private recurseLoadDataGroups(currentNode: Object,  parentGroup: DataGroup, currentStack: string, groupID: number,  value: number) {
+    private recurseLoadDataGroups(currentNode: Object,  parentGroup: DataGroup, currentStack: string, groupID: number, dimension: string,  value: number) {
         if(!currentNode["name"]) {
             throw new Error("Cannot parse a DataGroup that has no name!")
         }
@@ -116,17 +116,17 @@ class DataStore {
 
         // Parse sources and layers (if present)
         if(currentNode["sources"]) {
-            dataGroup.parseDataSources(currentNode["sources"], value);
+            dataGroup.parseDataSources(currentNode["sources"], dimension, value);
         }   1
         if(currentNode["layers"]) {
-            dataGroup.parseDataLayers(currentStack, currentNode["layers"], value);
+            dataGroup.parseDataLayers(currentStack, currentNode["layers"], dimension, value);
         }
 
         // Recurse into sub groups (if present)
         if(currentNode["groups"]) {
             for(let i = 0; i < currentNode["groups"].length; i++) {
                 let subNode = currentNode["groups"][i];
-                this.recurseLoadDataGroups(subNode, dataGroup, currentStack, i, value);
+                this.recurseLoadDataGroups(subNode, dataGroup, currentStack, i, dimension, value);
             }
         }
     }