+++ web/twa-vis-framework/library/CHANGELOG.md
@@ -9,12 +9,12 @@
 * Added functionality to scenario handler to fetch time series data from scenario endpoint
 * Added a `slider_handler` class to handle a slider for actions on time series data display queried from scenario
 * Specify actions for slider move in index.html file
-* Refresh and replot data with multiple layers in geoserver using `{wildcard}` in data.json (on slider change for example)
+* Refresh and replot data with multiple layers in geoserver using `{dimension}={value}` in data.json (on slider change for example)
 
 ## 4.6.0
 
 ### Features
-
+  
 * Functionality and styling for a navigation bar with keyloak logout integration. Used for CReDo but can be replicated for any KeyCloak server
 
 ### Bug Fixes